Added traceback, filehandling. Fixed a bug.
Traceback now shows the errors if the program crashes. This helps me to debug.

Filehandling makes it possible to remove or backup the .csv from within the program

Fixed a bug where the program would crash if you start it before a .csv exists and exit out without ever running start. And when you then restarted it would notice that the .csv was corrupted and if now you just said "Continue anyway" and THEN typed "STOP" the program would crash. Thanks to Tom for finding such a specific bug.